requirements.txt
dbt-core==1.7.0
dbt-postgres==1.7.0
psycopg2-binary==2.9.9
pandas==2.1.4
faker==21.0.0
streamlit==1.29.0
plotly==5.18.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
pytest==7.4.3
